MapMaker_장애물 비율 버그 수정
장애물 비율 계산시 100개를 넘어가면 일부 개수가 짤리는 문제 수정

+++ PacMan/PlayGame/GameMG/Map/Load/LoadMaker.cpp
@@ -118,7 +118,7 @@ void LoadMaker::_MakeAdditionalLoad(MapField map[][MAX_MAP_SIZE_X])
 	}
 
 	//남아있어야하는 장애물의 수
-	int temp = (curWallNum / 100) * _wallRatio;
+	int temp = ((float)curWallNum / 100.0f) * _wallRatio;
 	int wallNum = curWallNum - temp;
 
 	while (wallNum > 0)
